requirements.txt
# requirements.txt
streamlit>=1.22.0
streamlit-webrtc>=0.43.4
opencv-python>=4.6.0.66
fpdf2>=2.7.4
deepface>=0.0.79
av>=10.0.0
Pillow>=9.3.0